requirements.txt
absl-py==2.1.0
accelerate
aiohappyeyeballs
aiohttp
aiosignal
anndata==0.11.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_api_compat==1.9.1
arrow==1.3.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
Babel==2.15.0
beartype==0.19.0
bio==1.7.1
biopython==1.84
biothings-client==0.3.1
biotite==0.41.2
biotraj==1.2.2
bleach==6.1.0
blessed==1.20.0
cloudpathlib==0.20.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.1
deepspeed==0.15.4
defusedxml==0.7.1
dill==0.3.8
dm-tree==0.1.8
dnspython==2.6.1
drfp==0.3.6
easydict==1.13
einops==0.8.0
einx==0.3.0
expecttest==0.2.1
fair-esm==2.0.0
fastjsonschema==2.20.0
fonttools==4.53.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.3.1
gprofiler-official==1.0.0
gpustat==1.1.1
grpcio==1.64.1
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.2
hypothesis==6.100.1
ipykernel==6.29.4
isoduration==20.11.0
jaxtyping==0.2.36
joblib==1.4.2
json5==0.9.25
jsonpointer==2.1
kiwisolver==1.4.5
lark==1.1.9
legacy-api-wrap==1.4.1
llvmlite==0.43.0
lmdb==1.5.1
local-attention==1.9.15
loralib==0.1.2
Markdown==3.6
matplotlib==3.9.0
mistune==3.0.2
mkl-service==2.4.0
ml_collections==1.0.0
mpi4py==3.1.6
msgpack==1.1.0
msgpack-numpy==0.4.8
multidict==6.1.0
multiprocess==0.70.16
mygene==3.2.2
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
ninja==1.11.1.1
notebook_shim==0.2.4
numba==0.60.0
nvidia-ml-py==12.555.43
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.11.0
overrides==7.7.0
pandas==2.2.2
pandocfilters==1.5.1
patsy==1.0.1
pooch==1.8.2
prometheus_client==0.20.0
propcache==0.2.1
protobuf==4.25.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-etcd==0.4.5
python-json-logger==2.0.7
pyzmq==26.0.3
rdkit==2024.3.6
rdkit-pypi==2022.9.5
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rotary-embedding-torch==0.8.6
safetensors==0.4.5
scanpy==1.10.4
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
session-info==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
statsmodels==0.14.4
stdlib-list==0.11.0
svgwrite==1.4.3
taylor-series-linear-attention==0.1.12
tenacity==9.0.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.20.3
torch==2.3.0
torch-cluster@https://data.pyg.org/whl/torch-2.3.0%2Bcu121/torch_cluster-1.6.3%2Bpt23cu121-cp310-cp310-linux_x86_64.whl
torch-geometric==2.6.1
torch_scatter@https://data.pyg.org/whl/torch-2.3.0%2Bcu121/torch_scatter-2.1.2%2Bpt23cu121-cp310-cp310-linux_x86_64.whl
torchaudio==2.3.0
torchelastic==0.2.2
torchtext==0.18.0
torchvision==0.18.0
tornado==6.4.1
tqdm==4.67.1
transformers==4.46.2
Tree==0.2.4
triton==2.3.0
types-dataclasses==0.6.6
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
umap-learn==0.5.7
uri-template==1.3.0
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
xxhash==3.5.0
yarl==1.18.3
zstd==1.5.5.1
